requirements.txt
Flask==1.1.1line-bot-sdk==1.8.0beautifulsoup4==4.8.2requests==2.22.0sqlalchemy==1.3.13psycopg2==2.8.4